raise an AccessDenied exception in both known access denied scenarios
Previously we would see an `UnexpectedCloudflareResponseStatus(403)`
when Cloudflare rejected a request due to invalid credentials. The
API will also reject requests if the `X-Auth-Email` or `X-Auth-Key`
headers are not formatted as it expects, but those responses are
`400 Bad Request` with some custom error JSON (which resulted in an
empty stream returned by our client methods).

After this change, both types of rejections will be mapped to a new
`AccessDenied` exception.